backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.13

# core langchain packages with flexible ranges to allow groq compatibility
langchain>=0.2.14,<0.3.0
langchain-openai>=0.1.22,<0.2.0
langchain-groq>=1.1.2
groq>=0.37.1
langgraph==0.2.16

openai==1.40.6

python-dotenv==1.0.1

google-search-results==2.4.2   # SerpAPI

sendgrid==6.11.0
requests
sqlalchemy
pydantic==1.10.13
